requirements.txt
discord.py==2.3.2
googletrans==3.0.0
Requests==2.31.0
scrapetube==2.5.1
